refactor: reubicación de archivos de configuración y limpieza de estructura src
Simplificación de la arquitectura eliminando carpetas redundantes (po, lib, config).
Actualización de alias de TypeScript y resolución de conflictos de tipos en variables de entorno.
