--- src/config/env.ts
@@ -1,6 +1,0 @@
-export const CONFIG = {
-  BASE_URL: process.env.BASE_URL || 'https://practicesoftwaretesting.com/',
-  API_URL: process.env.API_URL || '',
-  CUSTOMER_EMAIL: process.env.CUSTOMER_EMAIL || '',
-  CUSTOMER_PASSWORD: process.env.CUSTOMER_PASSWORD || '',
-};
