+++ src/datafactories/register.ts
@@ -1,8 +1,7 @@
 import { expect, APIRequestContext } from '@playwright/test';
-import { CONFIG } from '@/config/env.js';
 
 export async function registerUser(request: APIRequestContext, user: Record<string, unknown>) {
-  const apiUrl = CONFIG.API_URL || 'https://api.practicesoftwaretesting.com';
+  const apiUrl = process.env.API_URL || 'https://api.practicesoftwaretesting.com';
   const response = await request.post(`${apiUrl}/users/register`, {
     data: user,
   });
