+++ src/pages/base.page.ts
@@ -1,5 +1,4 @@
 import { Page } from '@playwright/test';
-import { CONFIG } from '@/config/env.js';
 
 export default class BasePage {
   readonly page: Page;
@@ -9,8 +8,7 @@ export default class BasePage {
   }
 
   async goto(path: string) {
-    // Usa la variable de entorno BASE_URL si está definida, si no, usa la URL por defecto.
-    const baseUrl = CONFIG.BASE_URL;
+    const baseUrl = process.env.BASE_URL;
     await this.page.goto(`${baseUrl}/${path}`, { waitUntil: 'domcontentloaded' });
   }
 }
