+++ tsconfig.json
@@ -16,5 +16,5 @@
     "types": ["node", "@playwright/test"]
   },
   "include": ["src/**/*.ts", "playwright.config.ts"],
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules", "dist"]
 }
